move browser helper funcs to main plugin files
